refactor: move stack trace parsing logic to lib
The logic for generating truncated and readable stack traces can be
re-used and should be moved to lib.
